+++ src/applications/appeals/10182/tests/components/ContestableIssuesWidget.unit.spec.jsx
@@ -3,10 +3,10 @@ import { expect } from 'chai';
 import { mount } from 'enzyme';
 import sinon from 'sinon';
 
-import EligibleIssuesWidget from '../../components/EligibleIssuesWidget';
+import { ContestableIssuesWidget } from '../../components/ContestableIssuesWidget';
 import { SELECTED } from '../../constants';
 
-describe('<EligibleIssuesWidget>', () => {
+describe('<ContestableIssuesWidget>', () => {
   const getProps = ({
     review = false,
     submitted = false,
@@ -16,9 +16,8 @@ describe('<EligibleIssuesWidget>', () => {
     value: [
       { attributes: { ratingIssueSubjectText: 'issue-1' } },
       { attributes: { ratingIssueSubjectText: 'issue-2' } },
-      { attributes: { ratingIssueSubjectText: 'issue-3' } },
     ],
-    additionalIssues: [],
+    additionalIssues: [{ issue: 'issue-3' }],
     onChange,
     formContext: {
       onReviewPage: review,
@@ -29,8 +28,10 @@ describe('<EligibleIssuesWidget>', () => {
 
   it('should render a list of check boxes (IssueCard component)', () => {
     const props = getProps();
-    const wrapper = mount(<EligibleIssuesWidget {...props} />);
-    expect(wrapper.find('IssueCard').length).to.equal(props.value.length);
+    const wrapper = mount(<ContestableIssuesWidget {...props} />);
+    expect(wrapper.find('input[type="checkbox"]').length).to.equal(
+      props.value.length + props.additionalIssues.length,
+    );
     expect(
       wrapper
         .find('.widget-title')
@@ -39,10 +40,21 @@ describe('<EligibleIssuesWidget>', () => {
     ).to.equal('issue-1');
     wrapper.unmount();
   });
+  it('should render change link & remove button', () => {
+    const props = getProps();
+    const wrapper = mount(<ContestableIssuesWidget {...props} />);
+    const addLength = props.additionalIssues.length;
+    const link = wrapper.find('a.change-issue-link');
+    expect(link.length).to.equal(addLength);
+    expect(wrapper.find('button.remove-issue').length).to.equal(
+      props.additionalIssues.length,
+    );
+    wrapper.unmount();
+  });
 
   it('should not wrap the checkboxes in a fieldset', () => {
     const props = getProps();
-    const wrapper = mount(<EligibleIssuesWidget {...props} />);
+    const wrapper = mount(<ContestableIssuesWidget {...props} />);
     expect(wrapper.find('fieldset').length).to.equal(0);
     wrapper.unmount();
   });
@@ -50,7 +62,7 @@ describe('<EligibleIssuesWidget>', () => {
   it('should call onChange when the checkbox is toggled', () => {
     const onChange = sinon.spy();
     const props = getProps({ onChange });
-    const wrapper = mount(<EligibleIssuesWidget {...props} />);
+    const wrapper = mount(<ContestableIssuesWidget {...props} />);
     wrapper.find('.form-checkbox').forEach((element, index) => {
       onChange.reset();
 
@@ -85,26 +97,28 @@ describe('<EligibleIssuesWidget>', () => {
     });
     wrapper.unmount();
   });
-  it('should not show an error when submitted with no selections in eligible issues', () => {
+  it('should not show an error on submission with one selection', () => {
     const props = getProps({ submitted: true });
     const issues = [{ [SELECTED]: true }];
     const wrapper = mount(
-      <EligibleIssuesWidget {...props} additionalIssues={issues} />,
+      <ContestableIssuesWidget {...props} additionalIssues={issues} />,
     );
     expect(wrapper.find('.usa-input-error').length).to.equal(0);
     wrapper.unmount();
   });
 
-  it('should show a message when no issues found', () => {
-    const props = getProps({ review: true });
-    const wrapper = mount(<EligibleIssuesWidget {...props} value={[]} />);
-    expect(wrapper.find('dt').text()).to.contain('No eligible issues found');
+  it('should show an error when submitted with no selections', () => {
+    const props = getProps({ submitted: true });
+    const wrapper = mount(<ContestableIssuesWidget {...props} />);
+    expect(wrapper.find('va-alert h3').length).to.equal(1);
     wrapper.unmount();
   });
   it('should show a message when no issues selected on review page', () => {
     const props = getProps({ review: true });
-    const wrapper = mount(<EligibleIssuesWidget {...props} value={[{}]} />);
-    expect(wrapper.find('dt').text()).to.contain('No issues selected');
+    const wrapper = mount(<ContestableIssuesWidget {...props} />);
+    expect(wrapper.find('dt').text()).to.contain(
+      'at least one issue, so we can process your request',
+    );
     wrapper.unmount();
   });
 });
