+++ src/applications/appeals/10182/tests/containers/FormApp.unit.spec.jsx
@@ -233,7 +233,6 @@ describe('FormApp', () => {
       ],
     };
     const formData = {
-      'view:hasIssuesToAdd': true,
       contestableIssues: contestableIssues.issues,
       additionalIssues: [{ issue: 'other issue', [SELECTED]: true }],
       veteran: {
