+++ src/applications/appeals/10182/tests/fixtures/data/maximal-test.json
@@ -80,7 +80,6 @@
         "view:selected": false
       }
     ],
-    "view:hasIssuesToAdd": true,
     "additionalIssues": [
       {
         "issue": "right arm",
