+++ src/applications/appeals/10182/tests/utils/helpers.unit.spec.js
@@ -10,8 +10,6 @@ import {
   getIssueDate,
   getIssueNameAndDate,
   hasDuplicates,
-  showAddIssuesPage,
-  showAddIssueQuestion,
   isEmptyObject,
   setInitialEditMode,
   issuesNeedUpdating,
@@ -72,20 +70,8 @@ describe('getSelected & getSelectedCount', () => {
     ]);
     expect(getSelectedCount(data, data.additionalIssues)).to.eq(1);
   });
-  it('should not return selected additional issues when Veteran chooses not to include them', () => {
-    const data = {
-      'view:hasIssuesToAdd': false,
-      additionalIssues: [
-        { type: 'no', [SELECTED]: false },
-        { type: 'ok', [SELECTED]: true },
-      ],
-    };
-    expect(getSelected(data)).to.deep.equal([]);
-    expect(getSelectedCount(data, data.additionalIssues)).to.eq(0);
-  });
   it('should return selected additional issues', () => {
     const data = {
-      'view:hasIssuesToAdd': true,
       additionalIssues: [
         { type: 'no', [SELECTED]: false },
         { type: 'ok', [SELECTED]: true },
@@ -102,7 +88,6 @@ describe('getSelected & getSelectedCount', () => {
         { type: 'no1', [SELECTED]: false },
         { type: 'ok1', [SELECTED]: true },
       ],
-      'view:hasIssuesToAdd': true,
       additionalIssues: [
         { type: 'no2', [SELECTED]: false },
         { type: 'ok2', [SELECTED]: true },
@@ -208,73 +193,6 @@ describe('hasDuplicates', () => {
       ],
     });
     expect(result).to.be.true;
-  });
-});
-
-describe('showAddIssuesPage', () => {
-  it('should return true when no contestable issues selected', () => {
-    expect(showAddIssuesPage({})).to.be.true;
-    expect(showAddIssuesPage({ contestableIssues: [{}] })).to.be.true;
-  });
-  it('should return true when question is set to "yes", or no contestable issues selected', () => {
-    expect(showAddIssuesPage({ 'view:hasIssuesToAdd': true })).to.be.true;
-    expect(
-      showAddIssuesPage({
-        'view:hasIssuesToAdd': true,
-        contestableIssues: [{ [SELECTED]: true }],
-      }),
-    ).to.be.true;
-    expect(
-      showAddIssuesPage({
-        'view:hasIssuesToAdd': true,
-        contestableIssues: [{}],
-      }),
-    ).to.be.true;
-    expect(
-      showAddIssuesPage({
-        'view:hasIssuesToAdd': false,
-        contestableIssues: [{}],
-      }),
-    ).to.be.true;
-  });
-  it('should return true when nothing is selected, and past the issues pages', () => {
-    // probably unselected stuff on the review & submit page
-    expect(
-      showAddIssuesPage({
-        'view:hasIssuesToAdd': true,
-        contestableIssues: [{}],
-        additionalIssues: [{}],
-      }),
-    ).to.be.true;
-    expect(
-      showAddIssuesPage({
-        'view:hasIssuesToAdd': false,
-        boardReviewOption: 'foo', // we're past the issues page
-        contestableIssues: [{}],
-        additionalIssues: [{}],
-      }),
-    ).to.be.true;
-  });
-  it('should return false when "no" is chosen and there is a selected contestable issue', () => {
-    expect(
-      showAddIssuesPage({
-        'view:hasIssuesToAdd': false,
-        contestableIssues: [{ [SELECTED]: true }],
-      }),
-    ).to.be.false;
-  });
-});
-
-describe('showAddIssueQuestion', () => {
-  it('should show add issue question when contestable issues selected', () => {
-    expect(showAddIssueQuestion({ contestableIssues: [{ [SELECTED]: true }] }))
-      .to.be.true;
-  });
-  it('should not show add issue question when no issues or none selected', () => {
-    expect(showAddIssueQuestion({ contestableIssues: [] })).to.be.false;
-    expect(showAddIssueQuestion({ contestableIssues: [{}] })).to.be.false;
-    expect(showAddIssueQuestion({ contestableIssues: [{ [SELECTED]: false }] }))
-      .to.be.false;
   });
 });
