+++ src/applications/appeals/10182/tests/validations/issues.unit.spec.js
@@ -1,0 +1,154 @@
+import { expect } from 'chai';
+import sinon from 'sinon';
+
+import { getDate } from '../../utils/dates';
+import {
+  SELECTED,
+  MAX_SELECTIONS,
+  MAX_ISSUE_NAME_LENGTH,
+} from '../../constants';
+
+import {
+  selectionRequired,
+  uniqueIssue,
+  maxIssues,
+  missingIssueName,
+  maxNameLength,
+} from '../../validations/issues';
+
+const _ = null;
+
+describe('uniqueIssue', () => {
+  const contestableIssues = [
+    {
+      attributes: {
+        ratingIssueSubjectText: 'test',
+        approxDecisionDate: '2021-01-01',
+      },
+    },
+  ];
+
+  it('should not show an error when there are no issues', () => {
+    const errors = { addError: sinon.spy() };
+    uniqueIssue(errors, _, _, _, _, _, {});
+    expect(errors.addError.notCalled).to.be.true;
+  });
+  it('should not show an error when there are duplicate contestable issues', () => {
+    const errors = { addError: sinon.spy() };
+    uniqueIssue(errors, _, _, _, _, _, {
+      contestableIssues: [contestableIssues[0], contestableIssues[0]],
+    });
+    expect(errors.addError.notCalled).to.be.true;
+  });
+  it('should not show an error when there are no duplicate issues (only date differs)', () => {
+    const errors = { addError: sinon.spy() };
+    uniqueIssue(errors, _, _, _, _, _, {
+      contestableIssues,
+      additionalIssues: [{ issue: 'test', decisionDate: '2021-01-02' }],
+    });
+    expect(errors.addError.notCalled).to.be.true;
+  });
+  it('should show an error when there is a duplicate additional issue', () => {
+    const errors = { addError: sinon.spy() };
+    uniqueIssue(errors, _, _, _, _, _, {
+      contestableIssues,
+      additionalIssues: [{ issue: 'test', decisionDate: '2021-01-01' }],
+    });
+    expect(errors.addError.called).to.be.true;
+  });
+  it('should show an error when there are multiple duplicate additional issue', () => {
+    const errors = { addError: sinon.spy() };
+    uniqueIssue(errors, _, _, _, _, _, {
+      contestableIssues,
+      additionalIssues: [
+        { issue: 'test2', decisionDate: '2021-02-01' },
+        { issue: 'test2', decisionDate: '2021-02-01' },
+      ],
+    });
+    expect(errors.addError.called).to.be.true;
+  });
+});
+
+describe('maxIssues', () => {
+  it('should not show an error when the array length is less than max', () => {
+    const errors = { addError: sinon.spy() };
+    maxIssues(errors, _, {});
+    expect(errors.addError.called).to.be.false;
+  });
+  it('should show not show an error when the array length is greater than max', () => {
+    const errors = { addError: sinon.spy() };
+    const validDate = getDate({ offset: { months: -2 } });
+    const template = {
+      issue: 'x',
+      decisionDate: validDate,
+      [SELECTED]: true,
+    };
+    maxIssues(errors, _, _, _, _, _, {
+      contestableIssues: new Array(MAX_SELECTIONS).fill(template),
+      additionalIssues: [template],
+    });
+    expect(errors.addError.called).to.be.true;
+  });
+});
+
+describe('selectionRequired', () => {
+  const getData = (selectContested = false, selectAdditional = false) => ({
+    contestableIssues: [
+      {
+        attributes: {
+          ratingIssueSubjectText: 'test',
+          approxDecisionDate: '2021-01-01',
+        },
+        [SELECTED]: selectContested,
+      },
+    ],
+    additionalIssues: [
+      {
+        issue: 'test 2',
+        decisionDate: '2021-01-01',
+        [SELECTED]: selectAdditional,
+      },
+    ],
+  });
+  it('should show an error when no issues are selected', () => {
+    const errors = { addError: sinon.spy() };
+    selectionRequired(errors, _, getData());
+    expect(errors.addError.called).to.be.true;
+  });
+  it('should show not show an error when a contestable issue is selected', () => {
+    const errors = { addError: sinon.spy() };
+    selectionRequired(errors, _, getData(true));
+    expect(errors.addError.called).to.be.false;
+  });
+  it('should show not show an error when an additional issue is selected', () => {
+    const errors = { addError: sinon.spy() };
+    selectionRequired(errors, _, getData(false, true));
+    expect(errors.addError.called).to.be.false;
+  });
+});
+
+describe('missingIssueName', () => {
+  it('should show an error when a name is missing', () => {
+    const errors = { addError: sinon.spy() };
+    missingIssueName(errors);
+    expect(errors.addError.called).to.be.true;
+  });
+  it('should show an error when a name is missing', () => {
+    const errors = { addError: sinon.spy() };
+    missingIssueName(errors, 'test');
+    expect(errors.addError.called).to.be.false;
+  });
+});
+
+describe('maxNameLength', () => {
+  it('should show an error when a name is too long', () => {
+    const errors = { addError: sinon.spy() };
+    maxNameLength(errors, 'ab '.repeat(MAX_ISSUE_NAME_LENGTH / 2));
+    expect(errors.addError.called).to.be.true;
+  });
+  it('should show an error when a name is not too long', () => {
+    const errors = { addError: sinon.spy() };
+    maxNameLength(errors, 'test');
+    expect(errors.addError.called).to.be.false;
+  });
+});
